refactor: Move the BinaryLoader class out of the Windows folder since the platform is specified in the file name of the class
